backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
anthropic==0.28.0
PyPDF2==3.0.1
python-docx==1.1.2
diff-match-patch==20230430
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
pydantic==2.7.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
httpx==0.27.0
openai==1.30.1
